[BACKPORT 2025.2][#31746] DocDB: Unify filesystem abstractions between yb_util and rocksdb libraries
Summary:
Merges the duplicate filesystem classes that previously existed in both yb::
and rocksdb:: namespaces into a single canonical set in yb::. The rocksdb
namespace retains transitional `using` aliases so existing rocksdb call sites
keep compiling.

Classes merged:
- WritableFile and WritableFileWrapper: rocksdb-style API folded into
  yb::WritableFile (now inherits FileWithNameAndUniqueId). yb-only methods
  (PreAllocate, AppendSlices, SizeOnDisk) kept with sensible defaults. Size()
  is canonical; GetFileSize() becomes a non-virtual alias. Flush(FlushMode) is
  retained — its FLUSH_SYNC variant on PosixWritableFile uses
  sync_file_range(WAIT_AFTER), which is cheaper than Sync()'s fdatasync (skips
  metadata). A non-virtual Flush() wraps Flush(FLUSH_ASYNC).
- FileLock: moved from env.h to file_system.h alongside other filesystem
  abstractions; rocksdb gets a `using` alias.
- PosixFileLock: moved out of two anonymous namespaces in env_posix.cc files
  into file_system_posix.h.
- WritableFileImpl (memenv): the two copies replaced with a single
  yb::InMemoryWritableFile in file_system_mem.h, matching the existing
  InMemorySequentialFile / InMemoryRandomAccessFile naming.
- PosixWritableFile: both implementations merged into one in
  file_system_posix.{h,cc}. The merged class supports both yb-style
  (writev-based scatter-gather Append, pending_sync_ optimization,
  LOG_SLOW_EXECUTION timing, sync_on_close, SizeOnDisk, test simulation flags)
  and rocksdb-style features (Fsync, IsSyncThreadSafe, InvalidateCache,
  protected Allocate/RangeSync for PrepareWrite, GetUniqueId, tail hole-punching
  in Close). Two constructors keep both callers (yb and rocksdb factories)
  unchanged.
- DoSync and GetLogicalFileSize helpers moved from env_posix.cc to
  file_system_posix.{h,cc} so the merged PosixWritableFile can use them.
- macOS fallocate(mode=0) stub duplicated into file_system_posix.cc so
  PreAllocate works on both Linux and macOS.

Forward declarations of `class WritableFile;` and `class FileLock;` in the
rocksdb namespace (which would conflict with the new `using` aliases) removed
from several table/env headers. rocksdb_fwd.h forward-declares yb::WritableFile
and re-exports it via `using`.

Follow-up work (FileFactory consolidation, the rest of Env itself) tracked
separately.

Original commit: 621bd3f2e14de62a8595d1e819ea33f3be3c5237 / D53381

Test Plan: Jenkins

Reviewers: timur, arybochkin, hsunder, esheng

Reviewed By: esheng

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D53381
