+++ src/yb/rocksdb/db/db_test_util.h
@@ -303,7 +303,7 @@ class SpecialEnv : public EnvWrapper {
 #endif  // !(defined NDEBUG) || !defined(OS_WIN)
         return base_->Close();
       }
-      Status Flush() override { return base_->Flush(); }
+      Status Flush(FlushMode mode) override { return base_->Flush(mode); }
       Status Sync() override {
         ++env_->sync_counter_;
         while (env_->delay_sstable_sync_.load(std::memory_order_acquire)) {
@@ -317,6 +317,7 @@ class SpecialEnv : public EnvWrapper {
       yb::IOPriority GetIOPriority() override {
         return base_->GetIOPriority();
       }
+      uint64_t Size() const override { return base_->Size(); }
       const std::string& filename() const override { return base_->filename(); }
     };
     class ManifestFile : public WritableFile {
@@ -332,7 +333,7 @@ class SpecialEnv : public EnvWrapper {
       }
       Status Truncate(uint64_t size) override { return base_->Truncate(size); }
       Status Close() override { return base_->Close(); }
-      Status Flush() override { return base_->Flush(); }
+      Status Flush(FlushMode mode) override { return base_->Flush(mode); }
       Status Sync() override {
         ++env_->sync_counter_;
         if (env_->manifest_sync_error_.load(std::memory_order_acquire)) {
@@ -341,7 +342,7 @@ class SpecialEnv : public EnvWrapper {
           return base_->Sync();
         }
       }
-      uint64_t GetFileSize() override { return base_->GetFileSize(); }
+      uint64_t Size() const override { return base_->Size(); }
       const std::string& filename() const override { return base_->filename(); }
 
      private:
@@ -374,7 +375,7 @@ class SpecialEnv : public EnvWrapper {
       }
       Status Truncate(uint64_t size) override { return base_->Truncate(size); }
       Status Close() override { return base_->Close(); }
-      Status Flush() override { return base_->Flush(); }
+      Status Flush(FlushMode mode) override { return base_->Flush(mode); }
       Status Sync() override {
         ++env_->sync_counter_;
         return base_->Sync();
@@ -382,6 +383,7 @@ class SpecialEnv : public EnvWrapper {
       bool IsSyncThreadSafe() const override {
         return env_->is_wal_sync_thread_safe_.load();
       }
+      uint64_t Size() const override { return base_->Size(); }
       const std::string& filename() const override { return base_->filename(); }
 
      private:
