+++ src/yb/rocksdb/rocksdb_fwd.h
@@ -20,7 +20,13 @@
 #include "yb/util/clone_ptr.h"
 #include "yb/util/enums.h"
 
+namespace yb {
+class WritableFile;
+}
+
 namespace rocksdb {
+
+using yb::WritableFile;
 
 class Arena;
 class Cache;
@@ -39,7 +45,6 @@ class IteratorFilter;
 class TableReader;
 class UserFrontier;
 class UserFrontiers;
-class WritableFile;
 class WriteBatch;
 
 struct BlockBasedTableOptions;
