+++ src/yb/rocksdb/table/adaptive_table_factory.h
@@ -29,7 +29,6 @@ namespace rocksdb {
 
 struct EnvOptions;
 
-class WritableFile;
 class Table;
 class TableBuilder;
