+++ src/yb/rocksdb/table/block_based_table_reader.h
@@ -51,7 +51,6 @@ class InternalKeyComparator;
 class Iterator;
 class TableCache;
 class TableReader;
-class WritableFile;
 struct BlockBasedTableOptions;
 struct EnvOptions;
 struct ReadOptions;
