+++ src/yb/rocksdb/table/plain_table_factory.h
@@ -31,7 +31,6 @@ namespace rocksdb {
 
 struct EnvOptions;
 
-class WritableFile;
 class Table;
 class TableBuilder;
