+++ src/yb/rocksdb/table/plain_table_key_coding.h
@@ -27,7 +27,6 @@
 
 namespace rocksdb {
 
-class WritableFile;
 struct ParsedInternalKey;
 struct PlainTableReaderFileInfo;
 enum PlainTableEntryType : unsigned char;
