+++ src/yb/rocksdb/util/env_posix.cc
@@ -131,11 +131,7 @@ void SetFD_CLOEXEC(int fd, const EnvOptions* options) {
   }
 }
 
-class PosixFileLock : public FileLock {
- public:
-  int fd_;
-  std::string filename;
-};
+using yb::PosixFileLock;
 
 class PosixEnv : public Env {
  public:
@@ -701,7 +697,7 @@ class PosixRocksDBFileFactory : public RocksDBFileFactory {
         // disable mmap writes
         EnvOptions no_mmap_writes_options = options;
         no_mmap_writes_options.use_mmap_writes = false;
-        *result = std::make_unique<PosixWritableFile>(fname, fd, no_mmap_writes_options);
+        *result = std::make_unique<yb::PosixWritableFile>(fname, fd, no_mmap_writes_options);
       }
     }
     return s;
@@ -745,7 +741,7 @@ class PosixRocksDBFileFactory : public RocksDBFileFactory {
         EnvOptions no_mmap_writes_options = options;
         no_mmap_writes_options.use_mmap_writes = false;
 
-        *result = std::make_unique<PosixWritableFile>(fname, fd, no_mmap_writes_options);
+        *result = std::make_unique<yb::PosixWritableFile>(fname, fd, no_mmap_writes_options);
       }
     }
     return s;
