+++ src/yb/util/env_posix.cc
@@ -211,23 +211,6 @@ class ScopedFdCloser {
 #define STATUS_IO_ERROR(context, err_number) \
     STATUS_FROM_ERRNO_SPECIAL_EIO_HANDLING(context, err_number)
 
-static Status DoSync(int fd, const string& filename) {
-  ThreadRestrictions::AssertIOAllowed();
-  if (FLAGS_never_fsync) {
-    return Status::OK();
-  }
-  if (FLAGS_writable_file_use_fsync) {
-    if (fsync(fd) < 0) {
-      return STATUS_IO_ERROR(filename, errno);
-    }
-  } else {
-    if (fdatasync(fd) < 0) {
-      return STATUS_IO_ERROR(filename, errno);
-    }
-  }
-  return Status::OK();
-}
-
 static Status DoOpen(const string& filename, Env::CreateMode mode, int* fd, int extra_flags = 0) {
   ThreadRestrictions::AssertIOAllowed();
   int flags = O_RDWR;
@@ -253,16 +236,6 @@ static Status DoOpen(const string& filename, Env::CreateMode mode, int* fd, int 
   }
   *fd = f;
   return Status::OK();
-}
-
-// Get logical on-disk file size using file descriptor.
-Result<uint64_t> GetLogicalFileSize(int fd, const string& filename) {
-  ThreadRestrictions::AssertIOAllowed();
-  struct stat st;
-  if (fstat(fd, &st) == -1) {
-    return STATUS_IO_ERROR(filename, errno);
-  }
-  return st.st_size;
 }
 
 template <class Extractor>
@@ -321,230 +294,6 @@ Result<struct statvfs> GetFilesystemStats(const std::string& path) {
 
   return stat;
 }
-
-// Use non-memory mapped POSIX files to write data to a file.
-//
-// TODO (perf) investigate zeroing a pre-allocated allocated area in
-// order to further improve Sync() performance.
-class PosixWritableFile : public WritableFile {
- public:
-  PosixWritableFile(const std::string& fname, int fd, uint64_t file_size,
-                    bool sync_on_close)
-      : filename_(fname),
-        fd_(fd),
-        sync_on_close_(sync_on_close),
-        filesize_(file_size),
-        pre_allocated_size_(0),
-        pending_sync_(false) {}
-
-  ~PosixWritableFile() {
-    if (fd_ >= 0) {
-      WARN_NOT_OK(Close(), "Failed to close " + filename_);
-    }
-  }
-
-  Status Append(const Slice& data) override {
-    vector<Slice> data_vector;
-    data_vector.push_back(data);
-    return AppendVector(data_vector);
-  }
-
-  Status AppendSlices(const Slice* slices, size_t num) override {
-    ThreadRestrictions::AssertIOAllowed();
-    static const size_t kIovMaxElements = IOV_MAX;
-
-    Status s;
-    for (size_t i = 0; i < num && s.ok(); i += kIovMaxElements) {
-      size_t n = std::min(num - i, kIovMaxElements);
-      s = DoWritev(slices + i, n);
-    }
-
-    pending_sync_ = true;
-    return s;
-  }
-
-  Status PreAllocate(uint64_t size) override {
-    TRACE_EVENT1("io", "PosixWritableFile::PreAllocate", "path", filename_);
-    ThreadRestrictions::AssertIOAllowed();
-    uint64_t offset = std::max(filesize_, pre_allocated_size_);
-    if (PREDICT_FALSE(FLAGS_TEST_simulate_fs_without_fallocate)) {
-      YB_LOG_FIRST_N(WARNING, 1) << "Simulating a filesystem without fallocate() support";
-      return Status::OK();
-    }
-    if (PREDICT_FALSE(FLAGS_TEST_simulate_free_space_bytes >= 0 &&
-                      implicit_cast<uint64_t>(FLAGS_TEST_simulate_free_space_bytes) < size)) {
-      return STATUS_IO_ERROR(filename_, ENOSPC);
-    }
-    if (fallocate(fd_, 0, offset, size) < 0) {
-      if (errno == EOPNOTSUPP) {
-        YB_LOG_FIRST_N(WARNING, 1) << "The filesystem does not support fallocate().";
-      } else if (errno == ENOSYS) {
-        YB_LOG_FIRST_N(WARNING, 1) << "The kernel does not implement fallocate().";
-      } else {
-        return STATUS_IO_ERROR(filename_, errno);
-      }
-      // We don't want to modify pre_allocated_size_ since nothing was pre-allocated.
-      return Status::OK();
-    }
-    pre_allocated_size_ = offset + size;
-    return Status::OK();
-  }
-
-  Status Close() override {
-    TRACE_EVENT1("io", "PosixWritableFile::Close", "path", filename_);
-    ThreadRestrictions::AssertIOAllowed();
-    if (FLAGS_TEST_skip_file_close) {
-      return Status::OK();
-    }
-    Status s;
-
-    // size_on_disk can be either pre_allocated_size_ or pre_allocated_size_ from
-    // previous writer. If we've allocated more space than we used, truncate to the
-    // actual size of the file and perform Sync().
-    uint64_t size_on_disk = lseek(fd_, 0, SEEK_END);
-    if (size_on_disk < 0) {
-       return STATUS_IO_ERROR(filename_, errno);
-    }
-    if (filesize_ < size_on_disk) {
-      if (ftruncate(fd_, filesize_) < 0) {
-        s = STATUS_IO_ERROR(filename_, errno);
-        pending_sync_ = true;
-      }
-    }
-
-    if (sync_on_close_) {
-      Status sync_status = Sync();
-      if (!sync_status.ok()) {
-        LOG(WARNING) << "Unable to Sync " << filename_ << ": " << sync_status;
-        if (s.ok()) {
-          s = sync_status;
-        }
-      }
-    }
-
-    if (close(fd_) < 0) {
-      if (s.ok()) {
-        s = STATUS_IO_ERROR(filename_, errno);
-      }
-    }
-
-    fd_ = -1;
-    return s;
-  }
-
-  Status Flush(FlushMode mode) override {
-    TRACE_EVENT1("io", "PosixWritableFile::Flush", "path", filename_);
-    ThreadRestrictions::AssertIOAllowed();
-    if (FLAGS_never_fsync) {
-      return Status::OK();
-    }
-#if defined(__linux__)
-    int flags = SYNC_FILE_RANGE_WRITE;
-    if (mode == FLUSH_SYNC) {
-      flags |= SYNC_FILE_RANGE_WAIT_AFTER;
-    }
-    if (sync_file_range(fd_, 0, 0, flags) < 0) {
-      return STATUS_IO_ERROR(filename_, errno);
-    }
-#else
-    if (fsync(fd_) < 0) {
-      return STATUS_IO_ERROR(filename_, errno);
-    }
-#endif
-    return Status::OK();
-  }
-
-  Status Sync() override {
-    TRACE_EVENT1("io", "PosixWritableFile::Sync", "path", filename_);
-    ThreadRestrictions::AssertIOAllowed();
-    LOG_SLOW_EXECUTION(WARNING, 1000, Substitute("sync call for $0", filename_)) {
-      if (pending_sync_) {
-        pending_sync_ = false;
-        RETURN_NOT_OK(DoSync(fd_, filename_));
-      }
-    }
-    return Status::OK();
-  }
-
-  uint64_t Size() const override {
-    return filesize_;
-  }
-
-  Result<uint64_t> SizeOnDisk() const override {
-    TRACE_EVENT1("io", "PosixWritableFile::SizeOnDisk", "path", filename_);
-    return GetLogicalFileSize(fd_, filename_);
-  }
-
-  const string& filename() const override { return filename_; }
-
- protected:
-    void UnwrittenRemaining(struct iovec** remaining_iov, ssize_t written, int* remaining_count) {
-      size_t bytes_to_consume = written;
-      do {
-        if (bytes_to_consume >= (*remaining_iov)->iov_len) {
-          bytes_to_consume -= (*remaining_iov)->iov_len;
-          (*remaining_iov)++;
-          (*remaining_count)--;
-        } else {
-          (*remaining_iov)->iov_len -= bytes_to_consume;
-          (*remaining_iov)->iov_base =
-              static_cast<uint8_t*>((*remaining_iov)->iov_base) + bytes_to_consume;
-          bytes_to_consume = 0;
-        }
-      } while (bytes_to_consume > 0);
-    }
-
-    const std::string filename_;
-    int fd_;
-    bool sync_on_close_;
-    uint64_t filesize_;
-    uint64_t pre_allocated_size_;
-    std::atomic<bool> pending_sync_;
-
- private:
-  Status DoWritev(const Slice* slices, size_t n) {
-    ThreadRestrictions::AssertIOAllowed();
-    DCHECK_LE(n, IOV_MAX);
-
-    struct iovec iov[n];
-    ssize_t nbytes = 0;
-
-    for (size_t i = 0; i < n; ++i) {
-      const Slice& data = slices[i];
-      iov[i].iov_base = const_cast<uint8_t*>(data.data());
-      iov[i].iov_len = data.size();
-      nbytes += data.size();
-    }
-
-    struct iovec* remaining_iov = iov;
-    int remaining_count = narrow_cast<int>(n);
-    ssize_t total_written = 0;
-
-    while (remaining_count > 0) {
-      ssize_t written = writev(fd_, remaining_iov, remaining_count);
-      if (PREDICT_FALSE(written == -1)) {
-        if (errno == EINTR || errno == EAGAIN) {
-          continue;
-        }
-        return STATUS_IO_ERROR(filename_, errno);
-      }
-
-      TrackStackTrace(StackTraceTrackingGroup::kWriteIO, written);
-      UnwrittenRemaining(&remaining_iov, written, &remaining_count);
-      total_written += written;
-    }
-
-    filesize_ += total_written;
-
-    if (PREDICT_FALSE(total_written != nbytes)) {
-      return STATUS_FORMAT(
-          IOError, "writev error: expected to write $0 bytes, wrote $1 bytes instead",
-          nbytes, total_written);
-    }
-
-    return Status::OK();
-  }
-};
 
 #if defined(__linux__)
 class PosixDirectIOWritableFile final : public PosixWritableFile {
@@ -1000,12 +749,6 @@ static int LockOrUnlock(const std::string& fname,
   }
   return value;
 }
-
-class PosixFileLock : public FileLock {
- public:
-  int fd_;
-  std::string filename;
-};
 
 class PosixEnv : public Env {
  public:
