+++ src/yb/util/file_system_posix.h
@@ -15,6 +15,8 @@
 
 #include "yb/util/file_system.h"
 
+struct iovec;
+
 namespace yb {
 
 #if defined(__linux__)
@@ -44,6 +46,13 @@ class PosixSequentialFile : public SequentialFile {
 #endif
   int fd_;
   bool use_os_buffer_;
+};
+
+// Identifies a locked file on the POSIX filesystem.
+class PosixFileLock : public FileLock {
+ public:
+  int fd_;
+  std::string filename;
 };
 
 // pread() based random-access file.
@@ -78,46 +87,60 @@ class PosixRandomAccessFile : public RandomAccessFile {
   bool use_os_buffer_;
 };
 
-} // namespace yb
+// fdatasync (or fsync, when FLAGS_writable_file_use_fsync is set) the given fd.
+Status DoSync(int fd, const std::string& filename);
 
-namespace rocksdb {
-
-// TODO(unify_env): remove `using` statement once filesystem classes are fully merged into yb
-// namespace:
-using yb::FileSystemOptions;
+// Get logical on-disk file size using file descriptor.
+Result<uint64_t> GetLogicalFileSize(int fd, const std::string& filename);
 
 class PosixWritableFile : public WritableFile {
- private:
+ public:
+  PosixWritableFile(const std::string& fname, int fd, const FileSystemOptions& options,
+                    uint64_t initial_size = 0, bool sync_on_close = false);
+
+  // Convenience constructor for callers that don't carry FileSystemOptions.
+  PosixWritableFile(const std::string& fname, int fd, uint64_t initial_size, bool sync_on_close)
+      : PosixWritableFile(fname, fd, FileSystemOptions::kDefault, initial_size, sync_on_close) {}
+
+  ~PosixWritableFile();
+
+  Status Append(const Slice& data) override;
+  Status AppendSlices(const Slice* slices, size_t num) override;
+  Status PreAllocate(uint64_t size) override;
+  Status Truncate(uint64_t size) override;
+  Status Close() override;
+  Status Flush(FlushMode mode) override;
+  Status Sync() override;
+  Status Fsync() override;
+  bool IsSyncThreadSafe() const override;
+  uint64_t Size() const override { return filesize_; }
+  Result<uint64_t> SizeOnDisk() const override;
+  Status InvalidateCache(size_t offset, size_t length) override;
+#ifdef ROCKSDB_FALLOCATE_PRESENT
+  size_t GetUniqueId(char* id) const override;
+#endif
+  const std::string& filename() const override { return filename_; }
+
+ protected:
+  Status DoWritev(const Slice* slices, size_t n);
+  static void UnwrittenRemaining(
+      struct iovec** remaining_iov, ssize_t written, int* remaining_count);
+
+#ifdef ROCKSDB_FALLOCATE_PRESENT
+  Status Allocate(uint64_t offset, uint64_t len) override;
+  Status RangeSync(uint64_t offset, uint64_t nbytes) override;
+#endif
+
   const std::string filename_;
   int fd_;
+  bool sync_on_close_;
   uint64_t filesize_;
+  uint64_t pre_allocated_size_;
+  std::atomic<bool> pending_sync_;
 #ifdef ROCKSDB_FALLOCATE_PRESENT
   bool allow_fallocate_;
   bool fallocate_with_keep_size_;
 #endif
-
- public:
-  PosixWritableFile(const std::string& fname, int fd,
-                    const FileSystemOptions& options);
-  ~PosixWritableFile();
-
-  // Means Close() will properly take care of truncate
-  // and it does not need any additional information
-  Status Truncate(uint64_t size) override;
-  Status Close() override;
-  Status Append(const Slice& data) override;
-  Status Flush() override;
-  Status Sync() override;
-  Status Fsync() override;
-  uint64_t GetFileSize() override;
-  bool IsSyncThreadSafe() const override;
-  Status InvalidateCache(size_t offset, size_t length) override;
-#ifdef ROCKSDB_FALLOCATE_PRESENT
-  Status Allocate(uint64_t offset, uint64_t len) override;
-  Status RangeSync(uint64_t offset, uint64_t nbytes) override;
-  size_t GetUniqueId(char* id) const override;
-#endif
-  const std::string& filename() const override { return filename_; }
 };
 
-} // namespace rocksdb
+} // namespace yb
